Fix non-determinism when inlining
A problem has been fixed which would cause the inliner to act
differently during different compiles of the same code. This
non-determinism could cause hard-to-debug problems for inliner bugs only
happening sometimes.

The source of the non-determinism was caused by the fact that the order
in which functions were considered for inlining was determined based on
the order of iteration of a HashMap, which is non-deterministic and
changes between runs. This has been temporarily resolved by iterating
through functions in order of their symbol IDs, which should be
consistent across compiles.