Add a struct containing all analysis structures
Previously, it was necessary to have a separate structure for storing
data from each kind of analysis. For optimization groups that used many
of these structures, this meant that a reference to each would have to
be passed individually. By storing these together, a single reference
can be passed that allows access to all analysis structures.