Fix dominance analysis for start block
Previously, there was a bug in dominance analysis where the start block
was being considered to be dominated by all other blocks. This is
incorrect: the start block is always dominated only by itself. This
could cause false back edges to be detected, resulting in loop
optimizations operating on non-loops.