Implement tail call elimination
A new optimization has been added which performs tail call elimination.
When detecting that a tail call is made to the current function, this
optimization will turn that call and return sequence into a store to
registers representing the current function's parameters followed by a
jump back to the start of the function. This essentially turns the
recursive tail call into a loop.