Implement more constant multiplication simplifications
Previously, a simplification rule was added to transform multiplication
by a power of 2 into a left shift. However, there are also a number of
other simplifications that are almost always beneficial (though they do
result in more instructions): namely, multiplication by a negative power
of 2 and by constants 1 higher or lower than a power of 2 (or negative
power of 2). New rules have been added to perform these simplifications.