Add an AMD64 peephole to eliminate redundant jumps
Under some circumstances, it is possible to end up with jumps to a label
which immediately follows the jump. The main cause of this is the return
instruction, which always adds an unconditional jump to the epilogue. A
new peephole pattern has been added to eliminate these, possibly
eliminating the label itself in the process.