Improve logging facilities
Previously, a &mut std::io::Write was passed around for logging
purposes. This had several drawbacks, requiring a call to unwrap() after
every writeln! or write! and always evaluating the arguments to be
logged.

This has now been replaced with a new facility with the custom
log_writeln! and log_write! macros. Instead of using a &mut Write, a
&mut Log is used instead, which is a thin wrapper around an optional
&mut Write. In the event that no &mut Write was provided, the arguments
of log_writeln! or log_write! need not be evaluated.