Fix infinite loop in end instruction propagation
Due to a missed corner case, it was possible for end instruction
propagation to propagate the end instruction from a block to itself. The
only time this can happen is when the block is an empty infinite loop,
containing only a jump to itself. The fact that an end instruction had
been propagated would then trigger the optimizer to perform another pass
of the block merge group, as end instruction propagation can feed into
other optimizations (e.g. dead block elimination). The next pass would
then do the same thing, resulting in the optimizer entering an infinite
loop.

This has been corrected by preventing end instruction propagation from
propagating a jump to the block being propagated from, as that is never
a useful optimization.