Fix AMD64 codegen generating a load into rax for void returns
Previously, the AMD64 codegen would unconditionally generate a load of
the return value into the rax register. However, the ret instruction is
special in that it can take a register of type IlType::Void if no value
should be returned. Under these circumstances, no load of rax is needed,
as the return value can never actually be used. The codegen has thus
been taught not to emit the unnecessary load under these circumstnaces.