Implement register sizes in the AMD64 register allocator
Previously, the AMD64 register allocator would always use a mov qword
whenever spilling or unspilling virtual registers. While this works for
all current data types, it requires that 8 bytes be reserved for i32
values on the stack. By passing register type information through to the
register allocator, it is now possible to use appropriately-sized mov
instructions and to allocate the minimum necessary size for spill slots
on the stack.

Additionally, the AMD64 register allocator will now make use of the
AMD64 red zone when it is possible to do so.