Switch XSrc::Imm to use 32-bit immediates
Previously, XSrc::Imm in the AMD64 codegen would use a 64-bit integer
for its immediate. However, there is generally only one kind of
instruction that can accept a 64-bit immediate: a mov into a register.
Having XSrc::Imm use 64-bit immediates was resulting in a lot of 32-bit
bounds checks being required and only benefitted this one particular
instruction, so it has been removed. Instead, when required, a mov of a
64-bit immediate into a register will simply be added as a special case
instruction.