Add proper type resolution
Previously, resolution of type names was hardcoded to recognize types
named "i32" and "bool". Obviously, this cannot handle user-defined types
or generics and thus needs improvement.

This system has now been replaced by keeping track of types on the
SymRefTable and adding the builtin "i32" and "bool" types to the
top-level SymRefTable.