Remove remnants of function IDs
Previously, the frontend had a concept of functions having IDs that were
separate to their symbol IDs. Unfortunately, this turned out to be not
very useful, so the backend never made use of this information. In line
with this, all remnants of that idea have been purged from the frontend.