Remove Copy impl from IlType
Previously, IlType implemented the Copy trait since there was no reason
not to do so at the time. However, after consideration, this was not a
good idea since adding compound types at the IL level (to make
optimizations on them easier) will require that IlType carry data that
cannot be Copy.

To allow the addition of such features in the future without too many
problems, the Copy impl for IlType has been preemptively removed.