Remove extra log message during ilgen
Previously, the fact that the result of a block was being computed
during ilgen would be logged twice. This was due to an extra log message
in generate_block_il, which is redundant due to the same log message
being present in generate_expression_il. The extra log message has now
been removed from generate_block_il to reduce confusion.