Add a new i1 type to the IL
Previously, the only integer type in the IL was i32, which meant that
booleans were being represented as i32. This had some unfortunate
implications in the AMD64 codegen, as the instruction for setting a
register based on a flag will only overwrite the bottom byte of the
register. As a result, it was necessary to perform an extra AND of the
register with 1 to clear the remainder of the register.

Adding a special type for boolean values (i1) avoids the need to do
this, as the codegen can be assured that the undefined part of the
register will never be used. This could also enable some more complex
optimizations in the future.