Fix amd64 RA crash due to missing unspill
When the amd64 register allocator encounters a mov from one virtual
register to another where the source is about to become dead, it will
avoid emitting an actual mov by simply adjusting the register
allocator's internal state. However, the previous implementation of this
optimization would fail if the source register was currently spilled,
since the code assumed the source would already be in a register. This
has been corrected by ensuring that this case will correctly unspill
spilled registers.