Add type information to call IL instructions
Previously, the IL instruction for performing a call did not carry any
type information on it. The only way to gain such information would be
to look up the function symbol in the SymDefTable. This would cause
unnecessary extra work when optimizing and generating code for calls, so
a new field has been added to IlInstructionKind::Call which carries the
return type of the function being called.