Fix incorrect codegen on AMD64 for jnz and jz
Previously, the jnz and jz instructions were generating test sequences
using dwords. However, since the i1 type was introduced, these
instructions have accepted i1 instead of i32. While it was previously
correct to use a dword test, i1 is represented as a byte value in the
AMD64 codegen, so this instruction would test garbage in the upper 24
bits of the register. This has been corrected and jnz and jz
instructions now generate a byte test on AMD64.