Add a new sel instruction
Previously, the IL had no way of representing a branchless select
between two values based on the result of a comparison. To allow such
code to be represented, a new sel instruction has been added. This
instruction will read the i1 condition code of its first operand and
will move the value of either its second operand (if the condition was
true) or its third operand (if the condition was false) into its target
register.