Fix copy elision moving across interfering copy instructions
A missing check in the copy elision optimization was allowing
instructions to move across other copy instructions which would clobber
one of their operands. This was especially noticeable when tail call
elimination kicked in if the arguments of the next call used the
arguments from the previous call in their calculations. This has been
corrected by ensuring that the interference check is done for copy
instructions as well.