Fix single-element tuple type encoding in DWARF info
Previously, single-element tuple types were being encoded in the DWARF
debug info as a typedef, but had an extra unused field. This would cause
debug info appearing after them to be ignored incorrectly as it would
make it appear that the debug info was being ended prematurely. This has
been corrected and the extra field is no longer encoded.