Add SymId info to inline sites in the IlSpanMap
Previously, it was not possible to know what exact function had been
inlined at a particular inline call site since the SymId of the inlined
function was never kept around for later use. In order to emit better
DWARF debug info, knowing this information is necessary, so it is now
kept track of in the IlSpanMap.